misc/HtmlTestSuite/results/test-0521.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Grid Layout Test: Grid positioned items in auto-fit tracks and gaps</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez Garcia-Boente"/>
<link href="https://drafts.csswg.org/css-grid-1/#abspos" rel="help" title="9. Absolute Positioning"/>
<link href="https://drafts.csswg.org/css-grid-1/#auto-repeat" rel="help" title="7.2.2.2. Repeat-to-fill: auto-fill and auto-fit repetitions"/>
<link href="https://drafts.csswg.org/css-grid-1/#valdef-repeat-auto-fit" rel="help"/>
<meta content="This test checks that positioned items ignore collapsed gaps, both before and after the first in-flow item ." name="assert"/>
<link href="/css/support/grid.css" rel="stylesheet"/>
<style>
.container {
  width: 200px;
}
.grid {
  position: relative;
  grid: 10px / repeat(auto-fit, 30px);
  grid-gap: 5px;
}
span {
  background: blue;
}
.abs {
  position: absolute;
  top:0; right:0; bottom:0; left:0;
  background: pink;
}
</style>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
</head><body onload="checkLayout('.grid')">
<div id="log"/>
<br/>
<div class="container">
    <div class="grid">
        <span class="abs" data-expected-height="10" data-expected-width="30" style="grid-column: 1 / 5"/>
        <span data-expected-height="10" data-expected-width="30" style="grid-column: 2"/>
    </div>
</div>
</body>
</html>
